Unknown changes..

+++ Base20/Base20/src/interfaceCalc/AbstractProcessor.java
@@ -1,27 +1,27 @@
 package interfaceCalc;
 
 public abstract class AbstractProcessor implements Processorable {
-	 private int result;
-	 private int temp; 
+	 private long result;
+	 private long temp; 
 	 
 	 @Override
 	 public void printResult() {
 	  System.out.println("Result = " + result);
 	 }
 	 
-	 public void setResult(int r) {
+	 public void setResult(long r) {
 		result = r;
 	}
 	 
-	 public int getResult() {
+	 public long getResult() {
 		return result;
 	}
 	 
-	 public void setTemp(int tmp) {
+	 public void setTemp(long tmp) {
 		temp = tmp;
 	}
 	 
-	 public int getTemp() {
+	 public long getTemp() {
 		return temp;
 	}
 }
